@@ -20,15 +20,28 @@ package net.tridentsdk.server.world;
 import com.google.common.collect.ImmutableSet;
 import com.google.common.collect.Sets;
 import com.google.common.io.ByteStreams;
-import net.tridentsdk.Difficulty;
-import net.tridentsdk.GameMode;
-import net.tridentsdk.Position;
-import net.tridentsdk.Trident;
+import net.tridentsdk.*;
 import net.tridentsdk.base.Block;
 import net.tridentsdk.entity.Entity;
+import net.tridentsdk.entity.Projectile;
+import net.tridentsdk.entity.block.SlotProperties;
 import net.tridentsdk.entity.living.Player;
+import net.tridentsdk.entity.living.ProjectileLauncher;
+import net.tridentsdk.entity.traits.EntityProperties;
+import net.tridentsdk.entity.types.EntityType;
+import net.tridentsdk.entity.types.HorseType;
+import net.tridentsdk.entity.types.VillagerCareer;
+import net.tridentsdk.entity.types.VillagerProfession;
 import net.tridentsdk.factory.Factories;
 import net.tridentsdk.meta.nbt.*;
+import net.tridentsdk.server.entity.TridentDroppedItem;
+import net.tridentsdk.server.entity.TridentEntity;
+import net.tridentsdk.server.entity.TridentExpOrb;
+import net.tridentsdk.server.entity.TridentFirework;
+import net.tridentsdk.server.entity.block.*;
+import net.tridentsdk.server.entity.living.*;
+import net.tridentsdk.server.entity.projectile.*;
+import net.tridentsdk.server.entity.vehicle.*;
 import net.tridentsdk.server.packets.play.out.PacketPlayOutTimeUpdate;
 import net.tridentsdk.server.player.TridentPlayer;
 import net.tridentsdk.server.threads.ThreadsHandler;
@@ -36,15 +49,23 @@ import net.tridentsdk.util.TridentLogger;
 import net.tridentsdk.world.*;
 import net.tridentsdk.world.gen.ChunkAxisAlignedBoundingBox;
 
+import javax.annotation.concurrent.ThreadSafe;
 import java.io.*;
 import java.nio.file.Files;
 import java.nio.file.Paths;
 import java.util.Collection;
 import java.util.Set;
+import java.util.UUID;
 import java.util.concurrent.ThreadLocalRandom;
 import java.util.zip.GZIPInputStream;
 import java.util.zip.GZIPOutputStream;
 
+/**
+ * Represents a world on the server
+ *
+ * @author The TridentSDK Team
+ */
+@ThreadSafe
 public class TridentWorld implements World {
     private static final int SIZE = 1;
     private static final int MAX_HEIGHT = 255;
@@ -54,7 +75,7 @@ public class TridentWorld implements World {
     private final Set<Entity> entities = Factories.collect().createSet();
     private final String name;
     private final WorldLoader loader;
-    private final Position spawnLocation;
+    private final Position spawnPosition;
 
     private volatile long time;
     private volatile long existed;
@@ -74,13 +95,13 @@ public class TridentWorld implements World {
     private TridentWorld(String name, WorldLoader loader, boolean throwaway) {
         this.name = name;
         this.loader = loader;
-        this.spawnLocation = Position.create(this, 0, 0, 0);
+        this.spawnPosition = Position.create(this, 0, 0, 0);
     }
 
     TridentWorld(String name, WorldLoader loader) {
         this.name = name;
         this.loader = loader;
-        this.spawnLocation = Position.create(this, 0, 0, 0);
+        this.spawnPosition = Position.create(this, 0, 0, 0);
 
         TridentLogger.log("Starting to load " + name + "...");
 
@@ -116,9 +137,9 @@ public class TridentWorld implements World {
         }
 
         TridentLogger.log("Loading values of level.dat....");
-        spawnLocation.setX(((IntTag) level.getTag("SpawnX")).value());
-        spawnLocation.setY(((IntTag) level.getTag("SpawnY")).value() + 5);
-        spawnLocation.setZ(((IntTag) level.getTag("SpawnZ")).value());
+        spawnPosition.setX(((IntTag) level.getTag("SpawnX")).value());
+        spawnPosition.setY(((IntTag) level.getTag("SpawnY")).value() + 5);
+        spawnPosition.setZ(((IntTag) level.getTag("SpawnZ")).value());
 
         dimension = Dimension.OVERWORLD;
         // difficulty = Difficulty.difficultyOf(((IntTag) level.getTag("Difficulty")).value()); from tests does
@@ -152,8 +173,8 @@ public class TridentWorld implements World {
 
         TridentLogger.log("Loading spawn chunks...");
 
-        int centX = ((int) Math.floor(spawnLocation.x())) >> 4;
-        int centZ = ((int) Math.floor(spawnLocation.z())) >> 4;
+        int centX = ((int) Math.floor(spawnPosition.x())) >> 4;
+        int centZ = ((int) Math.floor(spawnPosition.z())) >> 4;
 
         for (int x = centX - 3; x <= centX + 3; x++) {
             for (int z = centZ - 3; z <= centZ + 3; z++) {
@@ -206,13 +227,13 @@ public class TridentWorld implements World {
 
             // TODO: load other values
 
-            world.spawnLocation.setX(0);
-            world.spawnLocation.setY(64);
-            world.spawnLocation.setZ(0);
+            world.spawnPosition.setX(0);
+            world.spawnPosition.setY(64);
+            world.spawnPosition.setZ(0);
 
             TridentLogger.log("Loading spawn chunks...");
-            int centX = ((int) Math.floor(world.spawnLocation.x())) >> 4;
-            int centZ = ((int) Math.floor(world.spawnLocation.z())) >> 4;
+            int centX = ((int) Math.floor(world.spawnPosition.x())) >> 4;
+            int centZ = ((int) Math.floor(world.spawnPosition.z())) >> 4;
 
             for (ChunkLocation location :
                     new ChunkAxisAlignedBoundingBox(ChunkLocation.create(centX - 7, centZ - 7),
@@ -231,52 +252,53 @@ public class TridentWorld implements World {
     }
 
     public void tick() {
-        ThreadsHandler.worldExecutor().execute(new Runnable() {
-            @Override
-            public void run() {
-                redstoneTick = !redstoneTick;
+        ThreadsHandler.worldExecutor().execute(() -> {
+            redstoneTick = !redstoneTick;
 
-                if (time >= 2400)
-                    time = 0;
-                if (time % 40 == 0)
-                    TridentPlayer.sendAll(new PacketPlayOutTimeUpdate().set("worldAge", existed).set("time", time));
+            if (time >= 2400)
+                time = 0;
+            if (time % 40 == 0)
+                TridentPlayer.sendAll(new PacketPlayOutTimeUpdate().set("worldAge", existed).set("time", time));
 
-                rainTime--;
-                thunderTime--;
+            rainTime--;
+            thunderTime--;
 
-                if (rainTime <= 0) {
-                    raining = !raining;
-                    rainTime = ThreadLocalRandom.current().nextInt();
-                }
+            if (rainTime <= 0) {
+                raining = !raining;
+                rainTime = ThreadLocalRandom.current().nextInt();
+            }
 
-                if (thunderTime <= 0) {
-                    thundering = !thundering;
-                    thunderTime = ThreadLocalRandom.current().nextInt();
-                }
+            if (thunderTime <= 0) {
+                thundering = !thundering;
+                thunderTime = ThreadLocalRandom.current().nextInt();
+            }
 
-                time++;
-                existed++;
+            for (Entity entity : entities) {
+                ((TridentEntity) entity).tick();
+            }
 
-                if (time % 150 == 0) {
-                    Set<ChunkLocation> set = Sets.newHashSet();
-                    for (Entity entity : entities) {
-                        if (entity instanceof Player) {
-                            Position pos = entity.location();
-                            int x = (int) pos.x() % 16;
-                            int z = (int) pos.z() % 16;
-                            int viewDist = Trident.config().getInt("view-distance", 7);
+            time++;
+            existed++;
 
-                            for (int i = x - viewDist; i < x + viewDist; i++) {
-                                for (int j = z - viewDist; j < z + viewDist; j++) {
-                                    set.add(ChunkLocation.create(i, j));
-                                }
+            if (time % Defaults.CHUNK_CLEAN_TICK_INTERVAL == 0) {
+                Set<ChunkLocation> set = Sets.newHashSet();
+                for (Entity entity : entities) {
+                    if (entity instanceof Player) {
+                        Position pos = entity.position();
+                        int x = (int) pos.x() / 16;
+                        int z = (int) pos.z() / 16;
+                        int viewDist = ((TridentPlayer) entity).viewDistance();
+
+                        for (int i = x - viewDist; i < x + viewDist; i++) {
+                            for (int j = z - viewDist; j < z + viewDist; j++) {
+                                set.add(ChunkLocation.create(i, j));
                             }
                         }
                     }
-
-                    loadedChunks.retain(set);
-                    set = null;
                 }
+
+                loadedChunks.retain(set);
+                set = null;
             }
         });
     }
@@ -299,9 +321,9 @@ public class TridentWorld implements World {
         TridentLogger.log("Saving " + name + "...");
         TridentLogger.log("Attempting to save level data...");
 
-        tag.addTag(new IntTag("SpawnX").setValue((int) spawnLocation.x()));
-        tag.addTag(new IntTag("SpawnY").setValue((int) spawnLocation.y()));
-        tag.addTag(new IntTag("SpawnZ").setValue((int) spawnLocation.z()));
+        tag.addTag(new IntTag("SpawnX").setValue((int) spawnPosition.x()));
+        tag.addTag(new IntTag("SpawnY").setValue((int) spawnPosition.y()));
+        tag.addTag(new IntTag("SpawnZ").setValue((int) spawnPosition.z()));
         tag.addTag(new DoubleTag("BorderSize").setValue(borderSize));
 
         tag.addTag(new ByteTag("Difficulty").setValue(difficulty.asByte()));
@@ -347,6 +369,20 @@ public class TridentWorld implements World {
                 TridentLogger.error(ex);
             }
         }
+    }
+
+    private Entity internalSpawn(Entity entity) {
+        ((TridentEntity) entity).spawn();
+        return addEntity(entity);
+    }
+
+    public Entity addEntity(Entity entity) {
+        this.entities.add(entity);
+        return entity;
+    }
+
+    public void removeEntity(Entity entity) {
+        this.entities.remove(entity);
     }
 
     @Override
@@ -408,7 +444,7 @@ public class TridentWorld implements World {
             // DEBUG =====
             TridentLogger.log("Generated chunk at (" + x + "," + z + ")");
             // =====
-            
+
             return chunk;
         }
 
@@ -458,8 +494,8 @@ public class TridentWorld implements World {
     }
 
     @Override
-    public Position spawnLocation() {
-        return spawnLocation;
+    public Position spawnPosition() {
+        return spawnPosition;
     }
 
     @Override
@@ -523,12 +559,171 @@ public class TridentWorld implements World {
     }
 
     @Override
-    public Set<Entity> entities() {
-        return ImmutableSet.copyOf(this.entities);
+    public Entity spawn(EntityType type, Position spawnPosition) {
+        switch (type) {
+            case NOT_IMPL:
+                throw new UnsupportedOperationException("Cannot spawn unimplemented entity");
+            case CREEPER:
+                return internalSpawn(new TridentCreeper(UUID.randomUUID(), spawnPosition));
+            case BAT:
+                return internalSpawn(new TridentBat(UUID.randomUUID(), spawnPosition));
+            case BLAZE:
+                return internalSpawn(new TridentBlaze(UUID.randomUUID(), spawnPosition));
+            case CHICKEN:
+                return internalSpawn(new TridentChicken(UUID.randomUUID(), spawnPosition));
+            case COW:
+                return internalSpawn(new TridentCow(UUID.randomUUID(), spawnPosition));
+            case ENDER_DRAGON:
+                return internalSpawn(new TridentEnderDragon(UUID.randomUUID(), spawnPosition));
+            case ENDERMAN:
+                return internalSpawn(new TridentEnderDragon(UUID.randomUUID(), spawnPosition));
+            case ENDERMITE:
+                return internalSpawn(new TridentEndermite(UUID.randomUUID(), spawnPosition));
+            case GHAST:
+                return internalSpawn(new TridentGhast(UUID.randomUUID(), spawnPosition));
+            case HORSE:
+                return internalSpawn(new TridentHorse(UUID.randomUUID(), spawnPosition, HorseType.HORSE));
+            case MAGMA_CUBE:
+                return internalSpawn(new TridentMagmaCube(UUID.randomUUID(), spawnPosition));
+            case MOOSHROOM:
+                return internalSpawn(new TridentMooshroom(UUID.randomUUID(), spawnPosition));
+            case OCELOT:
+                return internalSpawn(new TridentOcelot(UUID.randomUUID(), spawnPosition));
+            case PIG:
+                return internalSpawn(new TridentPig(UUID.randomUUID(), spawnPosition));
+            case GUARDIAN:
+                return internalSpawn(new TridentGuardian(UUID.randomUUID(), spawnPosition));
+            case PLAYER:
+                // Handle specially
+            case RABBIT:
+                return internalSpawn(new TridentRabbit(UUID.randomUUID(), spawnPosition));
+            case SHEEP:
+                return internalSpawn(new TridentSheep(UUID.randomUUID(), spawnPosition));
+            case SKELETON:
+                return internalSpawn(new TridentSkeleton(UUID.randomUUID(), spawnPosition));
+            case SLIME:
+                return internalSpawn(new TridentSlime(UUID.randomUUID(), spawnPosition));
+            case VILLAGER:
+                return internalSpawn(new TridentVillager(UUID.randomUUID(), spawnPosition,
+                        VillagerCareer.FARMER, VillagerProfession.FARMER));
+            case WITHER:
+                return internalSpawn(new TridentWither(UUID.randomUUID(), spawnPosition));
+            case WOLF:
+                return internalSpawn(new TridentWolf(UUID.randomUUID(), spawnPosition));
+            case ZOMBIE:
+                return internalSpawn(new TridentZombie(UUID.randomUUID(), spawnPosition));
+            case ARMOR_STAND:
+                return internalSpawn(new TridentArmorStand(UUID.randomUUID(), spawnPosition, new SlotProperties() {
+                }));
+            case FALLING_BLOCK:
+                return internalSpawn(new TridentFallingBlock(UUID.randomUUID(), spawnPosition));
+            case ITEM_FRAME:
+                return internalSpawn(new TridentItemFrame(UUID.randomUUID(), spawnPosition));
+            case PAINTING:
+                return internalSpawn(new TridentPainting(UUID.randomUUID(), spawnPosition));
+            case PRIMED_TNT:
+                return internalSpawn(new TridentPrimeTNT(UUID.randomUUID(), spawnPosition));
+            case ARROW:
+                return internalSpawn(new TridentArrow(UUID.randomUUID(), spawnPosition, new ProjectileLauncher() {
+                    @Override
+                    public <T extends Projectile> T launchProjectile(EntityProperties properties) {
+                        return null;
+                    }
+                }));
+            case EGG:
+                return internalSpawn(new TridentEgg(UUID.randomUUID(), spawnPosition, new ProjectileLauncher() {
+                    @Override
+                    public <T extends Projectile> T launchProjectile(EntityProperties properties) {
+                        return null;
+                    }
+                }));
+            case ENDER_PEARL:
+                return internalSpawn(new TridentEnderPearl(UUID.randomUUID(), spawnPosition, new ProjectileLauncher() {
+                    @Override
+                    public <T extends Projectile> T launchProjectile(EntityProperties properties) {
+                        return null;
+                    }
+                }));
+            case EXPERIENCE_BOTTLE:
+                return internalSpawn(new TridentExpBottle(UUID.randomUUID(), spawnPosition, new ProjectileLauncher() {
+                    @Override
+                    public <T extends Projectile> T launchProjectile(EntityProperties properties) {
+                        return null;
+                    }
+                }));
+            case FIREBALL:
+                return internalSpawn(new TridentFireball(UUID.randomUUID(), spawnPosition,
+                        new ProjectileLauncher() {
+                    @Override
+                    public <T extends Projectile> T launchProjectile(EntityProperties properties) {
+                        return null;
+                    }
+                }));
+            case FISH_HOOK:
+                return internalSpawn(new TridentFishHook(UUID.randomUUID(), spawnPosition, new ProjectileLauncher() {
+                    @Override
+                    public <T extends Projectile> T launchProjectile(EntityProperties properties) {
+                        return null;
+                    }
+                }));
+            case POTION:
+                return internalSpawn(new TridentPotion(UUID.randomUUID(), spawnPosition, new ProjectileLauncher() {
+                    @Override
+                    public <T extends Projectile> T launchProjectile(EntityProperties properties) {
+                        return null;
+                    }
+                }));
+            case SMALL_FIREBALL:
+                return internalSpawn(new TridentSmallFireball(UUID.randomUUID(), spawnPosition,
+                        new ProjectileLauncher() {
+                    @Override
+                    public <T extends Projectile> T launchProjectile(EntityProperties properties) {
+                        return null;
+                    }
+                }));
+            case SNOWBALL:
+                return internalSpawn(new TridentSnowball(UUID.randomUUID(), spawnPosition, new ProjectileLauncher() {
+                    @Override
+                    public <T extends Projectile> T launchProjectile(EntityProperties properties) {
+                        return null;
+                    }
+                }));
+            case WITHER_SKULL:
+                return internalSpawn(new TridentWitherSkull(UUID.randomUUID(), spawnPosition, new ProjectileLauncher() {
+                    @Override
+                    public <T extends Projectile> T launchProjectile(EntityProperties properties) {
+                        return null;
+                    }
+                }));
+            case BOAT:
+                return internalSpawn(new TridentBoat(UUID.randomUUID(), spawnPosition));
+            case COMMAND_MINECART:
+                return internalSpawn(new TridentCmdMinecart(UUID.randomUUID(), spawnPosition));
+            case FURNANCE_MINECART:
+                return internalSpawn(new TridentFurnaceMinecart(UUID.randomUUID(), spawnPosition));
+            case HOPPER_MINECART:
+                return internalSpawn(new TridentHopperMinecart(UUID.randomUUID(), spawnPosition));
+            case MINECART:
+                return internalSpawn(new TridentMinecart(UUID.randomUUID(), spawnPosition));
+            case SPAWNER_MINECART:
+                return internalSpawn(new TridentSpawnerMinecart(UUID.randomUUID(), spawnPosition));
+            case TNT_MINECART:
+                return internalSpawn(new TridentTntMinecart(UUID.randomUUID(), spawnPosition));
+            case ITEM:
+                return internalSpawn(new TridentDroppedItem(UUID.randomUUID(), spawnPosition));
+            case EXPERIENCE_ORB:
+                return internalSpawn(new TridentExpOrb(UUID.randomUUID(), spawnPosition));
+            case FIREWORK:
+                return internalSpawn(new TridentFirework(UUID.randomUUID(), spawnPosition));
+        }
+
+        // If it reaches here... that is really bad....
+        throw new UnsupportedOperationException("Cannot spawn that type of entity");
     }
 
-    public void addEntity(Entity entity) {
-        this.entities.add(entity);
+    @Override
+    public Set<Entity> entities() {
+        return ImmutableSet.copyOf(this.entities);
     }
 
     private static class PlayerFilter implements FilenameFilter {
