@@ -21,7 +21,6 @@ import com.google.code.tempusfugit.concurrency.ConcurrentRule;
 import com.google.code.tempusfugit.concurrency.RepeatingRule;
 import com.google.code.tempusfugit.concurrency.annotations.Concurrent;
 import com.google.code.tempusfugit.concurrency.annotations.Repeating;
-import io.netty.util.internal.chmv8.ConcurrentHashMapV8;
 import net.tridentsdk.AccessBridge;
 import net.tridentsdk.concurrent.ConcurrentCache;
 import net.tridentsdk.factory.CollectFactory;
@@ -32,6 +31,7 @@ import org.junit.Rule;
 import org.junit.Test;
 
 import java.util.concurrent.Callable;
+import java.util.concurrent.ConcurrentHashMap;
 import java.util.concurrent.ConcurrentMap;
 
 public class CacheTest extends AbstractTest {
@@ -40,7 +40,7 @@ public class CacheTest extends AbstractTest {
         AccessBridge.open().sendSuper(new CollectFactory() {
             @Override
             public <K, V> ConcurrentMap<K, V> createMap() {
-                return new ConcurrentHashMapV8<>();
+                return new ConcurrentHashMap<>();
             }
         });
         AccessBridge.open().sendSuper(ThreadsHandler.create());
@@ -58,12 +58,7 @@ public class CacheTest extends AbstractTest {
     @Repeating(repetition = 500)
     public void insert() {
         final Object object = new Object();
-        cache.retrieve(object, new Callable<Object>() {
-            @Override
-            public Object call() throws Exception {
-                return object;
-            }
-        });
+        cache.retrieve(object, () -> object);
         assertEquals(cache.remove(object), object);
     }
 
@@ -72,12 +67,7 @@ public class CacheTest extends AbstractTest {
     @Repeating(repetition = 500)
     public void insertNullValue() {
         final Object object = new Object();
-        cache.retrieve(object, new Callable<Object>() {
-            @Override
-            public Object call() throws Exception {
-                return null;
-            }
-        });
+        cache.retrieve(object, () -> null);
         assertNull(cache.remove(object));
     }
 
@@ -85,12 +75,7 @@ public class CacheTest extends AbstractTest {
     @Concurrent(count = 16)
     @Repeating(repetition = 500)
     public void insertNullKey() {
-        cache.retrieve(null, new Callable<Object>() {
-            @Override
-            public Object call() throws Exception {
-                return null;
-            }
-        });
+        cache.retrieve(null, () -> null);
     }
 
     @Test
@@ -98,12 +83,7 @@ public class CacheTest extends AbstractTest {
     @Repeating(repetition = 500)
     public void insertRemove() {
         final Object object = new Object();
-        cache.retrieve(object, new Callable<Object>() {
-            @Override
-            public Object call() throws Exception {
-                return object;
-            }
-        });
+        cache.retrieve(object, () -> object);
 
         assertEquals(cache.remove(object), object);
         assertNull(cache.remove(object));
