@@ -20,21 +20,13 @@ package net.tridentsdk.server.entity;
 import net.tridentsdk.Position;
 import net.tridentsdk.Trident;
 import net.tridentsdk.entity.Entity;
-import net.tridentsdk.entity.EntityBuilder;
-import net.tridentsdk.entity.ParameterValue;
 import net.tridentsdk.factory.ExecutorFactory;
-import net.tridentsdk.meta.nbt.CompoundTag;
-import net.tridentsdk.server.netty.ClientConnection;
-import net.tridentsdk.server.player.TridentPlayer;
 import net.tridentsdk.server.threads.ThreadsHandler;
-import net.tridentsdk.server.world.TridentWorld;
 import net.tridentsdk.util.TridentLogger;
-import net.tridentsdk.world.World;
 
 import java.lang.reflect.Constructor;
 import java.lang.reflect.InvocationTargetException;
 import java.util.UUID;
-import java.util.concurrent.Callable;
 
 /**
  * Builds an entity from initializer components and auto-spawns safely
@@ -43,62 +35,53 @@ import java.util.concurrent.Callable;
  *
  * @author The TridentSDK Team
  */
-public final class TridentEntityBuilder extends EntityBuilder {
+public final class EntityBuilder {
     private UUID uuid = UUID.randomUUID();
-    private Position spawn = Position.create(new Callable<World>() {
-        @Override
-        public World call() {
-            for (World world : Trident.worlds().values()) {
-                if (world.name().equals("world"))
-                    return world;
-            }
-            return null;
-        }
-    }.call(), 0, 0, 0);
-    private ExecutorFactory<Entity> executor;
+    private Position spawn = Position.create(Trident.worlds().get("world"), 0, 0, 0);
+    private ExecutorFactory executor;
     private boolean god;
     private Entity passenger;
     private String displayName;
     private boolean silent;
 
-    private TridentEntityBuilder() {
+    private EntityBuilder() {
     }
 
-    public static TridentEntityBuilder create() {
-        return new TridentEntityBuilder();
+    public static EntityBuilder create() {
+        return new EntityBuilder();
     }
 
-    public TridentEntityBuilder uuid(UUID uuid) {
+    public EntityBuilder uuid(UUID uuid) {
         this.uuid = uuid;
         return this;
     }
 
-    public TridentEntityBuilder spawn(Position spawn) {
+    public EntityBuilder spawn(Position spawn) {
         this.spawn = spawn;
         return this;
     }
 
-    public TridentEntityBuilder executor(ExecutorFactory<? extends Entity> executor) {
-        this.executor = (ExecutorFactory<Entity>) executor;
+    public EntityBuilder executor(ExecutorFactory executor) {
+        this.executor = executor;
         return this;
     }
 
-    public TridentEntityBuilder god(boolean god) {
+    public EntityBuilder god(boolean god) {
         this.god = god;
         return this;
     }
 
-    public TridentEntityBuilder passenger(Entity passenger) {
+    public EntityBuilder passenger(Entity passenger) {
         this.passenger = passenger;
         return this;
     }
 
-    public TridentEntityBuilder name(String displayName) {
+    public EntityBuilder name(String displayName) {
         this.displayName = displayName;
         return this;
     }
 
-    public TridentEntityBuilder silent(boolean silent) {
+    public EntityBuilder silent(boolean silent) {
         this.silent = silent;
         return this;
     }
@@ -136,22 +119,10 @@ public final class TridentEntityBuilder extends EntityBuilder {
         }
 
         TridentEntity entity = null;
-        if(entityType == TridentPlayer.class) {
-            entity = new TridentPlayer((CompoundTag)parameterValues[0].value(),
-                    (TridentWorld)parameterValues[1].value(),
-                    (ClientConnection)parameterValues[2].value());
-            entity.executor = executor != null ? executor : ThreadsHandler.entityExecutor();
-            entity.godMode = god;
-            entity.passenger = passenger;
-            entity.displayName = displayName;
-            entity.nameVisible = displayName != null;
-            entity.silent = silent;
-            entity.spawn();
-            return (T) entity;
-        }
         try {
             Constructor<? extends TridentEntity> constructor = (Constructor<? extends TridentEntity>)
-                    entityType.getConstructor(params);
+                    entityType.getDeclaredConstructor(params);
+            constructor.setAccessible(true);
             entity = constructor.newInstance(args);
             entity.executor = executor != null ? executor : ThreadsHandler.entityExecutor();
             entity.godMode = god;
@@ -166,5 +137,51 @@ public final class TridentEntityBuilder extends EntityBuilder {
         }
 
         return (T) entity;
+    }
+
+    /**
+     * Immutable parameter type and object value for dynamic constructor resolvation
+     *
+     * @param <T> the type for the parameter
+     * @author The TridentSDK Team
+     */
+    public static class ParameterValue<T> {
+        private final Class<T> c;
+        private final T value;
+
+        private ParameterValue(Class<T> c, T value) {
+            this.c = c;
+            this.value = value;
+        }
+
+        /**
+         * Creates a new parameter value
+         *
+         * @param c     the class type
+         * @param value the value of the parameter
+         * @param <T>   the type
+         * @return the new parameter value
+         */
+        public static <T> ParameterValue from(Class<T> c, T value) {
+            return new ParameterValue<>(c, value);
+        }
+
+        /**
+         * The class type for this parameter
+         *
+         * @return the parameter class type
+         */
+        public Class<T> clazz() {
+            return this.c;
+        }
+
+        /**
+         * The argument to be passed in for the parameter
+         *
+         * @return the value passed for the parameter
+         */
+        public T value() {
+            return this.value;
+        }
     }
 }
