@@ -19,7 +19,8 @@ package net.tridentsdk.server.entity.block;
 
 import net.tridentsdk.Position;
 import net.tridentsdk.base.Block;
-import net.tridentsdk.entity.EntityProperties;
+import net.tridentsdk.entity.traits.EntityProperties;
+import net.tridentsdk.entity.types.EntityType;
 import net.tridentsdk.entity.block.Painting;
 import net.tridentsdk.server.entity.TridentEntity;
 
@@ -36,17 +37,12 @@ public class TridentPainting extends TridentEntity implements Painting {
     }
 
     @Override
-    public boolean isNameVisible() {
-        return false;
-    }
-
-    @Override
-    public void applyProperties(EntityProperties properties) {
-
-    }
-
-    @Override
     public Block hangingBlock() {
         return null;
+    }
+
+    @Override
+    public EntityType type() {
+        return EntityType.PAINTING;
     }
 }
