@@ -19,7 +19,8 @@ package net.tridentsdk.server.entity.living;
 
 import net.tridentsdk.Position;
 import net.tridentsdk.entity.Entity;
-import net.tridentsdk.entity.EntityProperties;
+import net.tridentsdk.entity.traits.EntityProperties;
+import net.tridentsdk.entity.types.EntityType;
 import net.tridentsdk.entity.Projectile;
 import net.tridentsdk.entity.living.Endermite;
 import net.tridentsdk.entity.living.Player;
@@ -66,5 +67,10 @@ public class TridentEndermite extends TridentLivingEntity implements Endermite {
     @Override
     public <T extends Projectile> T launchProjectile(EntityProperties properties) {
         return null;
+    }
+
+    @Override
+    public EntityType type() {
+        return EntityType.ENDERMITE;
     }
 }
