@@ -18,10 +18,10 @@
 package net.tridentsdk.server.entity.projectile;
 
 import net.tridentsdk.Position;
-import net.tridentsdk.entity.EntityProperties;
+import net.tridentsdk.entity.traits.EntityProperties;
+import net.tridentsdk.entity.types.EntityType;
 import net.tridentsdk.entity.living.ProjectileLauncher;
 import net.tridentsdk.entity.projectile.Egg;
-import net.tridentsdk.server.entity.TridentProjectile;
 import net.tridentsdk.server.entity.living.TridentChicken;
 
 import java.util.UUID;
@@ -34,7 +34,7 @@ import java.util.concurrent.ThreadLocalRandom;
  */
 public class TridentEgg extends TridentProjectile implements Egg {
     /**
-     * Inherits constructor from {@link net.tridentsdk.server.entity.TridentProjectile}
+     * Inherits constructor from {@link TridentProjectile}
      */
     public TridentEgg(UUID uniqueId, Position spawnLocation, ProjectileLauncher source) {
         super(uniqueId, spawnLocation, source);
@@ -51,11 +51,16 @@ public class TridentEgg extends TridentProjectile implements Egg {
         if (chance <= 8) {
             if (chance == 1) {
                 for (int i = 0; i < 4; i++) {
-                    HANDLER.register(new TridentChicken(UUID.randomUUID(), this.location()));
+                    HANDLER.register(new TridentChicken(UUID.randomUUID(), this.position()));
                 }
             } else {
-                HANDLER.register(new TridentChicken(UUID.randomUUID(), this.location()));
+                HANDLER.register(new TridentChicken(UUID.randomUUID(), this.position()));
             }
         }
+    }
+
+    @Override
+    public EntityType type() {
+        return EntityType.EGG;
     }
 }
