@@ -18,10 +18,10 @@
 package net.tridentsdk.server.entity.projectile;
 
 import net.tridentsdk.Position;
-import net.tridentsdk.entity.EntityProperties;
+import net.tridentsdk.entity.traits.EntityProperties;
+import net.tridentsdk.entity.types.EntityType;
 import net.tridentsdk.entity.living.ProjectileLauncher;
 import net.tridentsdk.entity.projectile.WitherSkull;
-import net.tridentsdk.server.entity.TridentProjectile;
 
 import java.util.UUID;
 
@@ -36,5 +36,10 @@ public class TridentWitherSkull extends TridentProjectile implements WitherSkull
 
     @Override
     protected void hit() {
+    }
+
+    @Override
+    public EntityType type() {
+        return EntityType.WITHER_SKULL;
     }
 }
