@@ -20,7 +20,7 @@ package net.tridentsdk.server.packets.play.out;
 import io.netty.buffer.ByteBuf;
 import net.tridentsdk.Position;
 import net.tridentsdk.entity.Entity;
-import net.tridentsdk.entity.EntityType;
+import net.tridentsdk.entity.types.EntityType;
 import net.tridentsdk.server.netty.Codec;
 import net.tridentsdk.server.netty.packet.OutPacket;
 import net.tridentsdk.util.Vector;
@@ -37,7 +37,7 @@ public class PacketPlayOutSpawnObject extends OutPacket {
 
     @Override
     public void encode(ByteBuf buf) {
-        Position l = this.entity.location();
+        Position l = this.entity.position();
         Vector v = this.entity.velocity();
 
         Codec.writeVarInt32(buf, this.entityId);
